refactor package naming convention
