+++ src/main/java/org/example/Main.java
@@ -2,7 +2,7 @@ package org.example;
 
 import org.example.db.Database;
 import org.example.entity.User;
-import org.example.repository.UserRepositoryImp;
+import org.example.repository.userRepo.UserRepositoryImp;
 import org.example.services.UserService;
 
 
@@ -13,6 +13,7 @@ public class Main {
         UserService userService = new UserService(new UserRepositoryImp());
         userService.findById(1L);
         userService.getAllUsers();
+
         userService.addUser(User.builder()
                 .firstName("David")
                 .lastName("Davidsson")
