+++ src/main/java/org/example/repository/tripRepo/TripRepository.java
@@ -1,0 +1,15 @@
+package org.example.repository.tripRepo;
+
+import org.example.entity.Trip;
+import org.example.entity.User;
+
+import java.util.List;
+
+public interface TripRepository {
+    Trip get(Long id);
+    void add(Trip trip);
+    void update(Trip trip);
+    void remove(Trip trip);
+
+    List<Trip> getAllTrips();
+}
