+++ src/main/java/org/example/repository/tripRepo/TripRepositoryImp.java
@@ -1,0 +1,32 @@
+package org.example.repository.tripRepo;
+
+import org.example.entity.Trip;
+
+import java.util.List;
+
+public class TripRepositoryImp implements TripRepository {
+    @Override
+    public Trip get(Long id) {
+        return null;
+    }
+
+    @Override
+    public void add(Trip trip) {
+
+    }
+
+    @Override
+    public void update(Trip trip) {
+
+    }
+
+    @Override
+    public void remove(Trip trip) {
+
+    }
+
+    @Override
+    public List<Trip> getAllTrips() {
+        return null;
+    }
+}
